% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_info}
\alias{cluster_info}
\title{Fetch cluster information}
\usage{
cluster_info(cluster_name)
}
\arguments{
\item{cluster_name}{Name of the cluster (must exist)}
}
\description{
This function returns the row of information about a cluster
named \code{name_cluster} from the results in \code{gcloud container clusters list}.
}
