% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{docker_image}
\alias{docker_image}
\title{Build and push a docker image}
\usage{
docker_image(path, num_jobs)
}
\arguments{
\item{path}{Path to the image folder}

\item{num_jobs}{When run, the number of jobs spawned}
}
\value{
The path to the template directory
}
\description{
Once you create a docker directory for your script with
\code{\link[=docker_template]{docker_template()}}, this function is able to build it and push it to the
appropriate cloud diretory. Aditionally, it also creates all the job
yaml files so that the only remaining step is running the cluster
with \code{\link[=gcloud_run]{gcloud_run()}}.
}
\references{
\url{https://docs.docker.com/engine/reference/commandline/build/}
\url{https://cloud.google.com/container-registry/docs/pushing-and-pulling}
}
