% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{gcloud_cluster}
\alias{gcloud_cluster}
\title{Create kubernetes cluster}
\usage{
gcloud_cluster(name, machine_type = "g1-small", num_nodes = 3L,
  disk_size = "20GB", flags = list())
}
\arguments{
\item{name}{Name of the cluster}

\item{machine_type}{Type of the machine (\code{f1-micro}, \code{g1-small},
\code{n1-standard-1}, etc.)}

\item{num_nodes}{Number of nodes on cluster (must be an integer)}

\item{disk_size}{Startup disk size (defaults to \code{20GB})}

\item{flags}{List with more flags to use and their corresponding values}
}
\value{
If everything has gone as expected, \code{TRUE}
}
\description{
This function creates a kubernetes cluster on the current
project and zone via \code{gcloud container clusters create}. Note that
the defaults values for the parameters override the defaults of the
shell command.
}
\examples{
\dontrun{
create_cluster("r-cluster",
  disk_size = "100GB",
  flags = list("tags" = "k8s", "enable-autoupgrade" = "", "disk-type" = "pd-ssd")
)
}

}
\references{
\url{https://cloud.google.com/sdk/gcloud/reference/container/clusters/create}
}
