% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud.R
\name{gcloud_get_config}
\alias{gcloud_get_config}
\title{Get gcloud configuration}
\usage{
gcloud_get_config()
}
\value{
The string vector returned by the command (invisibly).
}
\description{
Prints to the console gcloud's current configuration by running
the \code{gcloud config list} command.
}
