% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud.R
\name{gcloud_set_config}
\alias{gcloud_set_config}
\title{Set gcloud configuration}
\usage{
gcloud_set_config(project = NULL, zone = NULL, region = NULL)
}
\arguments{
\item{project}{The desired \code{[PROJECT_ID]} (or \code{NULL} to skip)}

\item{zone}{The desired \code{[COMPUTE_ZONE]} (or \code{NULL} to skip)}

\item{region}{The desired \code{[COMPUTE_REGION]} (or \code{NULL} to skip)}
}
\value{
If everything has gone as expected, \code{TRUE}.
}
\description{
Sets gcloud's current project and zone configs by running
\code{gcloud config set project} and \code{gcloud config set compute/zone}.
}
