% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{has_docker}
\alias{has_docker}
\title{Check whether current user has docker setup}
\usage{
has_docker()
}
\description{
This function checks two things: if the curent user has docker
installed (via the \code{which docker} shell command) and if docker can be run
without sudo privileges (by testing the \code{docker info} command).
}
