% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud.R
\name{has_gcloud}
\alias{has_gcloud}
\title{Check whether current user has gcloud setup}
\usage{
has_gcloud()
}
\description{
This function checks if the curent user has gcloud installed
(via the \code{which gcloud} shell command).
}
