% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{kub_install_docker}
\alias{kub_install_docker}
\title{Install docker for the current user}
\usage{
kub_install_docker()
}
\value{
The path where docker was installed
}
\description{
This function can run two routines: full docker installation
if the user doesn't have it or correct privilege setup if the user runs
docker with sudo. The first method fetches and runs the script located
at \url{https://get.docker.com} and the second only runs
\code{sudo usermod -aG docker $USER}.
}
\section{Sudo}{

To install docker, this function requires sudo privileges. To do that, it
prompts the user for their password and maintains it until the end of
execution. All commands run this way are \strong{printed} to the console so
the user can know everything that's happening.
}

\references{
\url{https://docs.docker.com/install/linux/docker-ce/ubuntu/}
}
