% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud.R
\name{kub_install_gcloud}
\alias{kub_install_gcloud}
\title{Install gcloud for the current user}
\usage{
kub_install_gcloud()
}
\value{
The path where gcloud was installed
}
\description{
This function runs the gcloud installation routine via
\code{apt-get install google-cloud-sdk}. It also installs the \code{kubectl}
package for Kubernetes functionality. For more information, see the
"Getting started" vignette.
}
\section{Sudo}{

To install gcloud, this function requires sudo privileges. To do that, it
prompts the user for their password and maintains it until the end of
execution. All commands run this way are \strong{printed} to the console so
the user can know everything that's happening.
}

\references{
\url{https://cloud.google.com/sdk/docs/downloads-apt-get}
}
