% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{kub_kill_bucket}
\alias{kub_kill_bucket}
\title{Delete a storage bucket}
\usage{
kub_kill_bucket(name)
}
\arguments{
\item{name}{Name of the bucket}
}
\value{
The name of the bucket
}
\description{
This function deletes a storage bucket and all of its content
using the \code{gsutil rm} command.
}
\references{
\url{https://cloud.google.com/storage/docs/gsutil/commands/rm}
}
