% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{kub_kill_cluster}
\alias{kub_kill_cluster}
\title{Delete kubernetes cluster}
\usage{
kub_kill_cluster(name, zone = NULL)
}
\arguments{
\item{name}{Name of the cluster}

\item{zone}{Zone where the cluster was created (\code{NULL}, the current one, by
default)}
}
\value{
The name of the cluster
}
\description{
This function deletes a kubernetes cluster via
\code{gcloud container clusters delete}. Specify a zone if necessary.
}
\references{
\url{https://cloud.google.com/sdk/gcloud/reference/container/clusters/delete}
}
