% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{kub_kill_task}
\alias{kub_kill_task}
\title{Kill all kubernetes jobs and pods}
\usage{
kub_kill_task(path)
}
\arguments{
\item{path}{Path to task directory (if missing, defaults to the most recently
created task)}
}
\value{
If everything has gone as expected, \code{TRUE}
}
\description{
Using the command \code{kubectl delete}, deletes all jobs and pods
currently running in the cluster.
}
\references{
\url{https://kubernetes.io/docs/reference/generated/kubectl/kubectl-commands#delete}
}
