% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{kub_list_bucket}
\alias{kub_list_bucket}
\title{List contents of a gcloud bucket}
\usage{
kub_list_bucket(path, folder = "", recursive = FALSE)
}
\arguments{
\item{path}{Path to the image folder}

\item{folder}{List files in which directory from the bucket}

\item{recursive}{Whether to list files inside folders (\code{FALSE} by default)}
}
\value{
A character vector with file names
}
\description{
Using \code{gsutil ls}, this function is able to retrieve all
objects stored in a gclould bucket.
}
\references{
\url{https://cloud.google.com/storage/docs/gsutil/commands/ls}
}
