% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{kub_list_clusters}
\alias{kub_list_clusters}
\title{List all clusters available in current project}
\usage{
kub_list_clusters()
}
\value{
A table with clusters' information
}
\description{
Using the \code{gcloud container clusters list} command, this function
returns metadata about all clusters created in the current project.
}
\references{
\url{https://cloud.google.com/sdk/gcloud/reference/container/clusters/list}
}
