% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{kub_list_tasks}
\alias{kub_list_tasks}
\title{Get status of all currently running Kubernetes jobs}
\usage{
kub_list_tasks()
}
\value{
A table with jobs' status information
}
\description{
This function iterates on every available kubernetes context and
runs \code{kubectl get jobs} to get the statuses of their jobs.
}
\references{
\url{https://kubernetes.io/docs/reference/generated/kubectl/kubectl-commands#get}
}
