% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{kub_run_task}
\alias{kub_run_task}
\title{Run a job on a Kubernetes cluster}
\usage{
kub_run_task(path, cluster_name = NULL)
}
\arguments{
\item{path}{Path to the kuber directory}

\item{cluster_name}{Name of the cluster where to run the jobs (if \code{NULL},
the default, the name of the cluster set by \code{\link[=kub_init_task]{kub_init_task()}})}
}
\value{
The path to the kuber directory
}
\description{
Given a directory with an image build via \code{\link[=kub_push_task]{kub_push_task()}}, this
function assigns all jobs in the \code{/jobs} folder and runs them on the
associated Kubernetes cluster with the `.
}
\references{
\url{https://kubernetes.io/docs/reference/generated/kubectl/kubectl-commands#create}
}
