% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{kub_set_config}
\alias{kub_set_config}
\title{Set bucket and image information in a kuber directory}
\usage{
kub_set_config(path, parameters = list(cluster = NULL, bucket = NULL,
  image = NULL))
}
\arguments{
\item{path}{Path to task directory (if missing, defaults to the most recently
created task)}

\item{parameters}{A list with parameters to change: \code{cluster}, \code{bucket},
and/or \code{image} plus their respective values (or \code{NULL} to keep)}
}
\value{
If everything has gone as expected, \code{TRUE}
}
\description{
This function sets the names of the docker image and of the
gcloud bucket in \code{job-tmpl.yaml} at a kuber folder.
}
