% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{kuber_bucket}
\alias{kuber_bucket}
\title{Create a storage bucket}
\usage{
kuber_bucket(name)
}
\arguments{
\item{name}{Name of the bucket}
}
\description{
This function freates a default storage bucket using the
\code{gsutil mb} command.
}
\references{
\url{https://cloud.google.com/storage/docs/gsutil/commands/mb}
}
