% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{kuber_get_config}
\alias{kuber_get_config}
\title{Get bucket and image information from a kuber folder}
\usage{
kuber_get_config(path, what = "all", quiet = FALSE)
}
\arguments{
\item{path}{Path to the kuber directory}

\item{what}{What parameter to return: \code{all}, \code{cluster}, \code{location},
\code{region}, \code{num_nodes}, \code{bucket}, or \code{image}}

\item{quiet}{Whether to skip printing output (\code{FALSE} by default)}
}
\value{
A character vector with each parameter
}
\description{
This function extracts the names of all important kuber
parameters stored in the \code{.kuber} hidden file.
}
