% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{kuber_kill}
\alias{kuber_kill}
\title{Kill all kubernetes jobs and pods}
\usage{
kuber_kill()
}
\value{
TRUE
}
\description{
Using the command \code{kubectl delete}, deletes all jobs and pods
currently running in the cluster.
}
\references{
\url{https://kubernetes.io/docs/reference/generated/kubectl/kubectl-commands#delete}
}
