% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{kuber_list}
\alias{kuber_list}
\title{List contents of a gcloud bucket}
\usage{
kuber_list(path, recursive = FALSE)
}
\arguments{
\item{path}{Path to the image folder}

\item{recursive}{Whether to list files inside folders (\code{FALSE} by default)}
}
\value{
A character vector with file names
}
\description{
Using \code{gsutil ls}, this function is able to retrieve all
objects stored in a gclould bucket.
}
\references{
\url{https://cloud.google.com/storage/docs/gsutil/commands/ls}
}
