% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{kuber_pods}
\alias{kuber_pods}
\title{Get status of current Kubernetes pods}
\usage{
kuber_pods()
}
\description{
After running a group of jobs via \code{\link[=kuber_run]{kuber_run()}}, this function
runs \code{kubectl get pods} to get their statuses.
}
\references{
\url{https://kubernetes.io/docs/reference/generated/kubectl/kubectl-commands#get}
}
