% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{kuber_push}
\alias{kuber_push}
\title{Build and push the docker image of a kuber directory}
\usage{
kuber_push(path, num_jobs = NULL)
}
\arguments{
\item{path}{Path to the kuber directory}

\item{num_jobs}{When run, the number of jobs spawned (if \code{NULL}, the default,
\code{[NUM_NODES]} of the cluster)}
}
\value{
Path to the kuber directory
}
\description{
Once you create a docker directory for your script with
\code{\link[=kuber_init]{kuber_init()}}, this function is able to build it and push it to the
appropriate cloud diretory. Aditionally, it also creates all the job
yaml files so that the only remaining step is running the cluster
with \code{\link[=kuber_run]{kuber_run()}}.
}
\references{
\url{https://docs.docker.com/engine/reference/commandline/build/}
\url{https://cloud.google.com/container-registry/docs/pushing-and-pulling}
}
