% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{sys}
\alias{sys}
\title{Print and run command}
\usage{
sys(txt, ..., print = TRUE, ignore.stderr = FALSE)
}
\arguments{
\item{txt}{Text to print while running the command}

\item{...}{Command to be pasted and run}

\item{print}{Should anything be shown to the user?}

\item{ignore.stderr}{Whether to ignore stderr}
}
\value{
The result of the command as a character vector
}
\description{
Using \code{clisymbols} and \code{crayon}, this function prints a pretty
\code{usethis}-like step message with spinner and check-mark included.
}
