% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{todo}
\alias{todo}
\title{Print task for user}
\usage{
todo(...)
}
\arguments{
\item{...}{Text to be pasted and printed}
}
\description{
Using \code{clisymbols} and \code{crayon}, this function prints a pretty
taks list so the user knows what's the next step.
}
