% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gcloud_get_config}
\alias{gcloud_get_config}
\title{Get gcloud configuration}
\usage{
gcloud_get_config(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether to skip printing output (\code{FALSE} by default)}
}
\value{
The string vector returned by the command (invisibly)
}
\description{
Checks gcloud's current configuration by running the
\code{gcloud config list} command.
}
\references{
\url{https://cloud.google.com/sdk/gcloud/reference/config/list}
}
