% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{kuber_get_config}
\alias{kuber_get_config}
\title{Get bucket and image information from a kuber folder}
\usage{
kuber_get_config(path, quiet = FALSE)
}
\arguments{
\item{path}{Path to the kuber directory}

\item{quiet}{Whether to skip printing output (\code{FALSE} by default)}
}
\value{
The names of the bucket and image
}
\description{
This function extracts the names of the docker image and of the
gcloud bucket stored in \code{job-tmpl.yaml} at a kuber folder.
}
