% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{kuber_run}
\alias{kuber_run}
\title{Run a job on a Kubernetes cluster}
\usage{
kuber_run(path)
}
\arguments{
\item{path}{Path to the image folder}
}
\value{
The path to the kuber directory
}
\description{
Given a directory with an image build via \code{\link[=kuber_push]{kuber_push()}}, this
function assigns all jobs in the \code{/jobs} folder and runs them on the
associated Kubernetes cluster with the `.
}
\references{
\url{https://kubernetes.io/docs/reference/generated/kubectl/kubectl-commands#create}
}
