% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{kuber_run}
\alias{kuber_run}
\title{Run a job on a Kubernetes cluster}
\usage{
kuber_run(path, cluster_name = NULL)
}
\arguments{
\item{path}{Path to the kuber directory}

\item{cluster_name}{Name of the cluster where to run the jobs (if \code{NULL},
the default, the name of the cluster set by \code{\link[=kuber_init]{kuber_init()}})}
}
\value{
The path to the kuber directory
}
\description{
Given a directory with an image build via \code{\link[=kuber_push]{kuber_push()}}, this
function assigns all jobs in the \code{/jobs} folder and runs them on the
associated Kubernetes cluster with the `.
}
\references{
\url{https://kubernetes.io/docs/reference/generated/kubectl/kubectl-commands#create}
}
