% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{kuber_set_config}
\alias{kuber_set_config}
\title{Set bucket and image information in a kuber directory}
\usage{
kuber_set_config(path, bucket_name = NULL, image_name = NULL)
}
\arguments{
\item{path}{Path to the kuber directory}

\item{bucket_name}{Desired bucket name (or \code{NULL} to skip)}

\item{image_name}{Desired image name (or \code{NULL} to skip)}
}
\value{
If everything has gone as expected, \code{TRUE}
}
\description{
This function sets the names of the docker image and of the
gcloud bucket in \code{job-tmpl.yaml} at a kuber folder.
}
